resume_parser_app/requirements.txt
streamlit
pymupdf
python-docx
scikit-learn
pandas
plotly
pydantic